cleanup & consistency
