+++ Sources/ScopeKit/API/Types/CancellableOwner.swift
@@ -1,0 +1,13 @@
+import Combine
+import Foundation
+
+public protocol CancellableOwningWhileActive {
+    var whileActive: CancellableOwner { get }
+}
+
+public class CancellableOwner {
+    var externalCancellables = Set<AnyCancellable>()
+    public var cancellables: Set<AnyCancellable> {
+        externalCancellables
+    }
+}
