+++ Sources/ScopeKit/API/Types/ScopeIdentity.swift
@@ -1,0 +1,15 @@
+import Foundation
+
+public protocol ScopeIdentity: Identifiable, Hashable {
+    var id: UUID { get }
+}
+
+extension ScopeIdentity {
+    public static func == (lhs: Self, rhs: Self) -> Bool {
+        lhs.id == rhs.id
+    }
+
+    public func hash(into hasher: inout Hasher) {
+        hasher.combine(id)
+    }
+}
