+++ Sources/ScopeKit/API/Types/ScopeState.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public enum ScopeState {
+    case attached
+    case detached
+}
