+++ Sources/ScopeKit/Internal/HostComponent.swift
