+++ Sources/ScopeKit/Internal/InternalCancellable.swift
@@ -1,0 +1,21 @@
+import Combine
+import Foundation
+
+
+protocol InternalCancellable {
+    func cancel()
+}
+extension InternalCancellable {
+    func eraseToAnyCancellable() -> AnyCancellable {
+        AnyCancellable { self.cancel() }
+    }
+}
+
+extension CancellableOwner: InternalCancellable {
+    func cancel() {
+        externalCancellables.forEach { cancellable in
+            cancellable.cancel()
+        }
+        externalCancellables = Set<AnyCancellable>()
+    }
+}
