+++ Sources/ScopeKit/Internal/Startable.swift
@@ -2,3 +2,4 @@ import Combine
 import Foundation
 
 typealias Startable = () -> AnyCancellable
+typealias Callable = () -> ()
