+++ Sources/ScopeKit/Internal/Weak.swift
